requirements.txt
fastapi
uvicorn
pydantic
openai-whisper
ffmpeg-python


